python/pyproject.toml
[build-system]
requires = ["setuptools>=72", "wheel", "setuptools_scm[toml]>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "etos_api"
dynamic = ["version"]
description = "ETOS API"
authors = [{name = "Tobias Persson", email = "tobias.persson@axis.com"}]
license = { text  = "Apache License, Version 2.0" }
readme = "README.rst"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License"
]
requires-python = ">=3.9"
dependencies = [
    "etos_lib==4.4.2",
    "etcd3gw~=2.3",
    "uvicorn~=0.22",
    "fastapi~=0.109.1",
    "aiohttp[speedups]~=3.8",
    "gql[requests]~=3.4",
    "httpx~=0.24",
    "kubernetes~=26.1",
    "sse-starlette~=1.6",
    "opentelemetry-api~=1.21",
    "opentelemetry-exporter-otlp~=1.21",
    "opentelemetry-instrumentation-fastapi==0.46b0",
    "opentelemetry-sdk~=1.21",
    "jsontas~=1.4",
    "packageurl-python~=0.11",
    "cryptography>=42.0.4,<43.0.0",
]

[project.urls]
Documentation = "https://etos.readthedocs.io/"
Homepage = "https://github.com/eiffel-community/etos-api"
Repository = "https://github.com/eiffel-community/etos-api"

[project.optional-dependencies]
testing = ["pytest", "pytest-cov"]

[tool.build_sphinx]
source_dir = "docs"
build_dir = "build/sphinx"

[tool.devpi.upload]
no-vcs = 1
formats = "bdist_wheel"

[tool.flake8]
exclude = [".tox", "build", "dist", ".eggs", "docs/conf.py"]

[tool.pytest.ini_options]
addopts = "--cov etos_api --cov-report term-missing --verbose"
norecursedirs = ["dist", "build", ".tox"]
testpaths = ["tests"]

[tool.setuptools_scm]
root = ".."

[tool.setuptools.packages]
find = { where = ["src"], exclude = ["tests"] }
